Béta autominer

+++ src/fr/entasia/skycore/events/AMEvents.java
@@ -7,13 +7,16 @@ import fr.entasia.skycore.objs.AutoMiner;
 import fr.entasia.skycore.others.tasks.AutoMinerTask;
 import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.entity.ArmorStand;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.player.PlayerArmorStandManipulateEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
+import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.scheduler.BukkitRunnable;
 
 import java.util.ArrayList;
 
@@ -33,7 +36,7 @@ public class AMEvents implements Listener {
 
 	@EventHandler
 	public void onClick(PlayerInteractEvent e){
-		if(e.getAction()== Action.RIGHT_CLICK_BLOCK&&e.getClickedBlock().getType()== Material.HOPPER&&e.getPlayer().isSneaking()){
+		if(e.getHand()== EquipmentSlot.HAND&&e.getAction()== Action.RIGHT_CLICK_BLOCK&&e.getClickedBlock().getType()== Material.HOPPER&&e.getPlayer().isSneaking()){
 			ItemStack item = e.getPlayer().getInventory().getItemInMainHand();
 			if(pickaxes.contains(item.getType())){
 
@@ -45,24 +48,35 @@ public class AMEvents implements Listener {
 					e.getPlayer().sendMessage("§cCette île à déja atteint son maximum de mineur automatiques ! (" + AutoMiner.MAX + ")");
 				}else{
 					e.setCancelled(true);
+					System.out.println("cancelled");
 					for(AutoMiner am : is.autominers){
-						if(e.getClickedBlock()==am.hopper){
+						if(e.getClickedBlock().equals(am.hopper)){
 							e.getPlayer().getInventory().setItemInMainHand(am.pickaxe);
 							am.pickaxe = item;
+							for(ArmorStand as : am.armorStands){
+								as.setItemInHand(am.pickaxe);
+							}
 							return;
 						}
 					}
 
-					e.getPlayer().getInventory().setItemInMainHand(new ItemStack(Material.AIR));
-					AutoMiner am = new AutoMiner(e.getClickedBlock(), item);
-					AutoMiner.deleteByBlock(e.getClickedBlock());
-					am.spawn();
+					new BukkitRunnable() {
+						@Override
+						public void run() {
+							if(pickaxes.contains(item.getType())) {
+								e.getPlayer().getInventory().setItemInMainHand(new ItemStack(Material.AIR));
+								AutoMiner am = new AutoMiner(e.getClickedBlock(), item);
+								AutoMiner.deleteByBlock(e.getClickedBlock());
+								am.spawn();
 
-					AutoMinerTask.miners.add(am);
-					is.autominers.add(am);
-					Location loc = am.hopper.getLocation();
-					Main.sqlite.fastUpdate("INSERT INTO autominers (is_x, is_z, x, y, z, world, item) VALUES (?, ?, ?, ?, ?, ?, ?)",
-							is.isid.x, is.isid.z, loc.getBlockX(), loc.getBlockY(), loc.getBlockZ(), loc.getWorld().getName(), Serialization.serialiseItem(item));
+								AutoMinerTask.miners.add(am);
+								is.autominers.add(am);
+								Location loc = am.hopper.getLocation();
+								Main.sqlite.fastUpdate("INSERT INTO autominers (is_x, is_z, x, y, z, world, item) VALUES (?, ?, ?, ?, ?, ?, ?)",
+										is.isid.x, is.isid.z, loc.getBlockX(), loc.getBlockY(), loc.getBlockZ(), loc.getWorld().getName(), Serialization.serialiseItem(item));
+							}
+						}
+					}.runTaskLater(Main.main, 1);
 				}
 			}
 		}
@@ -74,7 +88,7 @@ public class AMEvents implements Listener {
 			BaseIsland is = BaseAPI.getIsland(CooManager.getIslandID(e.getBlock().getLocation()));
 			if(is==null)return;
 			for(AutoMiner am : is.autominers){
-				if(e.getBlock()==am.hopper){
+				if(e.getBlock().equals(am.hopper)){
 					am.delete();
 					am.ore.getWorld().dropItem(am.ore.getLocation(), am.pickaxe);
 				}
